reorganize FFI and environment files
