--- wasm/ffi.go
@@ -1,76 +1,0 @@
-package wasm
-
-import (
-	"github.com/pkg/errors"
-)
-
-/*
- In order to allow "easy" communication of data across the FFI barrier (outbound Go -> WASM and inbound WASM -> Go), hivew provides
- an FFI API. Functions exported from a WASM module can be easily called by Go code via the Wasmer instance exports, but returning data
- to the host Go code is not quite as straightforward.
-
- In order to accomplish this, hivew internally keeps a set of "environments" in a singleton package var (`environments` below).
- Each environment is a container that includes the WASM module bytes, and a set of WASM instances (runtimes) to execute said module.
- The envionment object has an index referencing its place in the singleton array, and each instance has an index referencing its position within
- the environment's instance array.
-
- When a WASM function calls one of the FFI API functions, it includes the `env_index` and `inst_index` values that were provided at the beginning
- of job execution, which allows hivew to look up the [env][instance] and send the result on the appropriate result channel. This is needed due to
- the way Go makes functions available on the FFI using CGO.
-*/
-
-/////////////////////////////////////////////////////////////////////////////
-// below is the wasm glue code used to manipulate wasm instance memory     //
-// this requires a set of functions to be available within the wasm module //
-// - allocate                                                              //
-// - deallocate                                                            //
-/////////////////////////////////////////////////////////////////////////////
-
-func (w *wasmInstance) readMemory(pointer int32, size int32) []byte {
-	data := w.wasmerInst.Memory.Data()[pointer:]
-	result := make([]byte, size)
-
-	for index := 0; int32(index) < size; index++ {
-		result[index] = data[index]
-	}
-
-	return result
-}
-
-func (w *wasmInstance) writeMemory(data []byte) (int32, error) {
-	lengthOfInput := len(data)
-
-	allocate := w.wasmerInst.Exports["allocate"]
-	if allocate == nil {
-		return -1, errors.New("missing required FFI function: allocate")
-	}
-
-	// Allocate memory for the input, and get a pointer to it.
-	allocateResult, err := allocate(lengthOfInput)
-	if err != nil {
-		return -1, errors.Wrap(err, "failed to call allocate")
-	}
-
-	pointer := allocateResult.ToI32()
-
-	w.writeMemoryAtLocation(pointer, data)
-
-	return pointer, nil
-}
-
-func (w *wasmInstance) writeMemoryAtLocation(pointer int32, data []byte) {
-	lengthOfInput := len(data)
-
-	// Write the input into the memory.
-	memory := w.wasmerInst.Memory.Data()[pointer:]
-
-	for index := 0; index < lengthOfInput; index++ {
-		memory[index] = data[index]
-	}
-}
-
-func (w *wasmInstance) deallocate(pointer int32, length int) {
-	dealloc := w.wasmerInst.Exports["deallocate"]
-
-	dealloc(pointer, length)
-}
